requirements.txt
groq==0.2.4
python-dotenv==1.0.1
